refactor: move editable table siderbar into separate folder
